requirements.txt
--extra-index-url https://download.pytorch.org/whl/cpu
PyPDF2==3.0.1
pycryptodome==3.18.0
openai==1.55.2
tiktoken==0.8.0
pdfminer.six==20221105
pgvector==0.1.8
sqlalchemy==2.0.17
psycopg2-binary==2.9.9
tenacity==8.2.2
azure-storage-blob==12.17.0
langdetect==1.0.9
numpy==1.25.0
azure-messaging-webpubsubservice==1.0.1
python-pptx>=0.6.22
openpyxl==3.1.2
pandas==2.0.3
lxml==4.9.3
xlrd==2.0.1
torch==2.1.1+cpu
tqdm==4.64.0

#chromadb>=0.3.23
python-doctr==0.8.0
#git+https://github.com/facebookresearch/detectron2.git@b7ff946
httpx==0.27.0
jdeskew==0.2.2
timm==0.9.11
fasttext-wheel==0.9.2
pdfplumber==0.9.0
deepdoctection==0.27
torchvision==0.16.1+cpu
azure-storage-queue==12.9.0
azure-identity
requests== 2.28.0
#pillow==9.5.0
langchain==0.3.9
langchain-core==0.3.21
langchain-community==0.3.8
langchain-openai==0.2.10
langgraph==0.2.56
Werkzeug==0.16.1
python-dotenv==1.0.1
pycountry==23.12.11
matplotlib<=3.8.4
pydantic==2.7.4
html2text==2024.2.26
alembic==1.13.2
mistral_common~=1.5.1
google-search-results==2.4.2
httpx==0.27.0
beautifulsoup4==4.12.3
hnswlib==0.8.0
scikit-learn==1.5.1
PyYAML==6.0
infinity_client==0.0.70
Unidecode==1.3.8
